requirements.txt
pyjnius==1.3.0
wrapt==1.13.3